% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_stack_downloads.R
\name{gmr_stack_urls}
\alias{gmr_stack_urls}
\title{Return the urls for all registered GMR-Gal4 stacks}
\usage{

  gmr_stack_urls(gmr_url = "http://flybrain.mrc-lmb.cam.ac.uk/vfb/jfrc/fl/reformatted-quant/")
}
\arguments{
\item{gmr_url}{The URL of the directory containing all registered GMR images
for download.}
}
\value{
A character vector of URLs each named by their GMR ID
}
\description{
Return the urls for all registered GMR-Gal4 stacks
}
\details{
You probably won't want to use this directly
}
\examples{
u=gmr_stack_urls()
length(u)
}

