% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_stack.R
\name{vfb_3dbrowser_url}
\alias{vfb_3dbrowser_url}
\title{Make a URL for a set of VFB ids to be displayed in the 3D browser}
\usage{
vfb_3dbrowser_url(ids, template = "VFBt_001",
  server = getOption("vfbr.server.gepetto"))
}
\arguments{
\item{ids}{A character vector of IDs}

\item{template}{VFB id of the template brain}

\item{server}{The base url of the server}
}
\value{
character vector to open 3D browser
}
\description{
Make a URL for a set of VFB ids to be displayed in the 3D browser
}
\examples{
# some flycircuit neurons
ids=c("VFB_00004657","VFB_00023207","VFB_00023120","VFB_00022264")
u=vfb_3dbrowser_url(ids)

fcids=c("VGlut-F-000304", "VGlut-F-200278", "fru-F-200121", "TH-F-300016")
u2=vfb_3dbrowser_url(vfb_tovfbids(fcids))
\dontrun{
browseURL(u)
browseURL(u2)
}
}

