% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfb_extended_query.R
\name{vfb_autocomplete_query}
\alias{vfb_autocomplete_query}
\title{Replicate the VFB site's autocomplete query}
\usage{
vfb_autocomplete_query(query, ...)
}
\arguments{
\item{query}{A character vector containing one or more queries - see details.
Wildcards (\code{?*}) can be used, and have a similar meaning to their use
in the shell. Queries are case insensitive.}

\item{...}{Additional arguments passed to \code{\link{vfb_solr_query}}. You
can use this e.g. to set the number of returned \code{rows}.}
}
\description{
Replicate the VFB site's autocomplete query
}
\details{
This is a rather non-selective query that emphasises exact matches and terms
in the BRAINNAME defining key ontology components by returning them in the
first rows of the result list.
}
\seealso{
\code{\link{vfb_synonym_query}}
}

