% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfbids.R
\name{vfb_fromvfbids}
\alias{vfb_fromvfbids}
\alias{vfb_tovfbids}
\title{Convert between VFB and external identifiers}
\usage{
vfb_fromvfbids(vfbids, ..., mustWork = NA)

vfb_tovfbids(ids, fixed = TRUE, ...)
}
\arguments{
\item{vfbids}{One or more vfb identifiers. Solr wildcards can also be used
when \code{fixed=FALSE}- see examples and \code{\link{vfb_solr_query}}}

\item{...}{Additional arguments passed to \code{\link{vfb_solr_query}}.}

\item{mustWork}{logical: if \code{TRUE} then an error is given if there are
missing results; if \code{NA} then a warning; if \code{FALSE} then there will be
no message but missing values will still be denoted by \code{NA} values.}

\item{ids}{One or more external identifiers. Solr wildcards can also be used
when \code{fixed=FALSE}- see examples and \code{\link{vfb_solr_query}}}

\item{fixed}{Whether to insist on exact matches. When \code{fixed=FALSE} solr
wildcards can be used.}
}
\value{
A character vector of vfb ids (for \code{vfb_fromvfbids}) or external
  identifiers (for \code{vfb_tovfbids}) in the same order as the input query.
  Missing values will be denoted by \code{NA}s when \code{fixed=TRUE}. When
  \code{fixed=FALSE} and a wild-card search returns no results a character
  vector of length 0 will be returned.
}
\description{
Convert between VFB and external identifiers
}
\details{
Note that these functions are not appropriate for fetching more than
  ~ 200 identifiers individually specified identifiers. If a large number of
  identifiers are required, it is a much better idea to fetch all ids and
  then filter locally.
}
\examples{
# some flycircuit ids
fcids=c("VGlut-F-000304", "VGlut-F-200278", "fru-F-200121", "TH-F-300016")
vfbids=vfb_tovfbids(fcids)
vfb_fromvfbids(vfbids)

# make up a fake id for testing - will give a warning
\donttest{
vfb_fromvfbids(c(vfbids, "VFB_10013392"))
}
# Some GMR GAL4 lines
gmrs=c('93D09', '87F10')
vfb_tovfbids(sprintf("GMR_\%s*", gmrs), fixed=FALSE)
}
\seealso{
\code{\link{gmr_vfbid}}, \code{\link{vfb_solr_query}}, which powers
  the underlying queries.
}
