% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vfb_query.R
\name{vfb_generic_query}
\alias{vfb_generic_query}
\title{Generic query against VFB API}
\usage{
vfb_generic_query(query, path = "do/jsonQuery.html?json=",
  server = getOption("vfbr.server"), parse.json = TRUE, ...)
}
\arguments{
\item{query}{A key-value list specifying the query}

\item{path}{The path on the server containing the query page}

\item{server}{The base url of the server}

\item{parse.json}{Whether or no to parse the response (default: TRUE)}

\item{...}{additional arguments passed to}
}
\description{
This function epects an R list or vector describing a query and constructs an
appropriate query url embedding a JSON query, GETs the server response and
(by default) parses the JSON result.
}
\examples{
vfb_generic_query(list(query_type="descendant_class", query="FBbt:00003679"))

vfb_generic_query(list(query_type="individuals", query="FBbt:00003679"))
}

