% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_nblast.R
\name{vfb_nblast}
\alias{vfb_nblast}
\title{Query vfb opencpu server to obtain pre-computed nblast scores}
\usage{
vfb_nblast(query, n = 50, ...)
}
\arguments{
\item{query}{A single flycircuit identifier}

\item{n}{The number of hits to return}

\item{...}{Additional arguments to be included in the POST}
}
\value{
a data.frame in descending score order with columns \itemize{

  \item id Flycircuit Neuron name

  \item score NBLAST score

  }
}
\description{
Query vfb opencpu server to obtain pre-computed nblast scores
}
\examples{
vfb_nblast('fru-M-200266', n=10)
}

