% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfb_query.R
\name{vfb_owl_query}
\alias{vfb_owl_query}
\title{Generic query against VFB OWL API}
\usage{
vfb_owl_query(query, path = "do/jsonQuery.html?json=",
  server = getOption("vfbr.server"), parse.json = TRUE, ...)
}
\arguments{
\item{query}{A key-value list specifying the query}

\item{path}{The path on the server containing the query page}

\item{server}{The base url of the server}

\item{parse.json}{Whether or no to parse the response (default: TRUE)}

\item{...}{additional arguments passed to
\code{jsonlite::\link[jsonlite]{fromJSON}}}
}
\description{
This function expects an R list or vector describing a query and
  constructs an appropriate query url embedding a JSON query, GETs the server
  response and (by default) parses the JSON result.
}
\details{
Note that the VFB OWL query endpoint by default wraps all results
  inside a JSON result object called \code{results}. When
  \code{parse.json=TRUE}, the returned results will be unwrapped to remove
  this outer layer.
}
\examples{
\dontrun{
# query for descendant classes of Fan-Shaped Body
vfb_owl_query(list(query_type="descendant_class", query="FBbt:00003679"))

# query for individual neurons overlapping with Fan-Shaped Body
neurondf=vfb_owl_query(list(query_type="individuals", query="FBbt:00003679"))
# show the first few rows of the returned data.frame
head(neurondf)
}
}
\seealso{
\code{\link[jsonlite]{fromJSON}}
}
