% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_query.R
\name{vfb_solr_query}
\alias{vfb_solr_query}
\title{Title}
\usage{
vfb_solr_query(query, path = "search/select?wt=json&df=short_form",
  fields = "label+short_form", sort = "score+desc", rows = 30L,
  server = getOption("vfbr.server"), parse.json = TRUE, ...)
}
\arguments{
\item{query}{A key-value list specifying the query}

\item{path}{The path on the server containing the query page}

\item{fields}{Which fields to return (+delimited). A value of \code{""}
implies all fields.}

\item{sort}{Character vector naming one or more fields (+ delimited) to use
for sorting the results.}

\item{rows}{Maximum number of rows to return}

\item{server}{The base url of the server}

\item{parse.json}{Whether or no to parse the response (default: TRUE)}

\item{...}{additional arguments passed to
\code{jsonlite::\link[jsonlite]{fromJSON}}}
}
\value{
When \code{parse.json=TRUE}, a list containing the parsed response
  with the main information contained in \code{rval$response$docs}. Otherwise
  an \code{httr::response} object
}
\description{
Title
}
\examples{
# Find VFB ids matching a given GMR line
vfb_solr_query("fq=VFB_*&q=label:GMR_10A07*")
\dontrun{
# VFB id for all GMRs
vfb_solr_query("fq=VFB_*&q=label:GMR_*")
}
}
\seealso{
\code{\link[httr]{response}}
}

