% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfb_stack_downloads.R
\name{vfb_stack_url_from_vfbid}
\alias{vfb_stack_url_from_vfbid}
\title{Find generic URL to image stack for download from VFB based on VFB ID}
\usage{
vfb_stack_url_from_vfbid(x)
}
\arguments{
\item{x}{A VFB id (either numeric or character)}
}
\value{
A character vector containing download URLs
}
\description{
Find generic URL to image stack for download from VFB based on VFB ID
}
\details{
This function expects the VFB id which has been designated for e.g.
  a GAL4 line image. You can find these on the VFB webpages as
  \code{VFB_XXXXXXXX}. In contrast to \code{\link{gmr_stack_urls_for_ids}}, this
  should work for the image data associated with any VFB id.
}
\examples{
\donttest{
vfb_stack_url_from_vfbid("VFB_00029638")
vfb_stack_url_from_vfbid(29638)

# find ids for a VT line
vtids=vfb_solr_query(filterquery="VFB_*",query="label:VT017929*")$short_form
# and then get the download URLs
vfb_stack_url_from_vfbid(vtids)

# find ids for a GMR line and then get the download URL
vfb_stack_url_from_vfbid(gmr_vfbid("13E10"))
}
}
\seealso{
\code{\link{gmr_stack_urls_for_ids}}
}
