% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_stack_downloads.R
\name{download_gmr_stacks}
\alias{download_gmr_stacks}
\title{Download the registered GMR stacks for given ids}
\usage{
download_gmr_stacks(ids, download.dir = getOption("vfbr.stack.downloads"),
  Force = FALSE, ...)
}
\arguments{
\item{ids}{Character vector of GMR ids specified in any way}

\item{download.dir}{The download directory}

\item{Force}{Whether to force the download.}

\item{...}{Additional arguments passed to \code{\link{download.file}}}
}
\value{
Named character vector of paths to downloaded files
}
\description{
Download the registered GMR stacks for given ids
}
\details{
Note that the downloading tries to be a little bit clever, by
  \enumerate{

  \item wrapping the \code{download.file} call in a try expression in case it
  fails

  \item memoising the \code{download.file} call so if it is called with the
  same url and destination value it will not re-download the file.

  }

  If memoisation gives you unexpected behaviour, you can set
  \code{Force=TRUE}.
}
\examples{
\dontrun{
stacks=download_gmr_stacks(c('9A09', '95H11'))
stacks

# on a mac this will open in the default app (e.g. Fiji/ImageJ)
open_stack<-function(x) system(paste("open", paste(shQuote(x), collapse = " ")))
open_stack(stacks)
# open the downloads directory in the Finder
system(paste("open -R", shQuote(getOption('vfbr.stack.downloads'))))
}
}

