% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfbids.R
\name{gmr_vfbid}
\alias{gmr_vfbid}
\title{Convert GMR identifiers to VFB ids}
\usage{
gmr_vfbid(ids)
}
\arguments{
\item{ids}{A character vector of GMR Gal4 ids}
}
\description{
Note that this function currently is vectorised and operates by pulling a
list of all GMR ids on VFB. Therefore if you have 50 (or 1000) GMR lines to
query, you should run the query in a single call to \code{gmr_vfbid}.
}
\examples{
gmr_vfbid("11H09")
}
\seealso{
\code{\link{vfb_fromvfbids}}
}
