% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfbids.R
\name{vfb_fromvfbids}
\alias{vfb_fromvfbids}
\alias{vfb_tovfbids}
\title{Convert between VFB and external identifiers}
\usage{
vfb_fromvfbids(vfbids, ..., mustWork = NA)

vfb_tovfbids(ids, synonyms = TRUE, fixed = TRUE, ...)
}
\arguments{
\item{vfbids}{One or more vfb identifiers. Solr wildcards can also be used
when \code{fixed=FALSE}- see examples and \code{\link{vfb_solr_query}}}

\item{...}{Additional arguments passed to \code{\link{vfb_solr_query}}.}

\item{mustWork}{logical: if \code{TRUE} then an error is given if there are
missing results; if \code{NA} then a warning; if \code{FALSE} then there will be
no message but missing values will still be denoted by \code{NA} values.}

\item{ids}{One or more external identifiers. Solr wildcards can also be used
when \code{fixed=FALSE}- see examples and \code{\link{vfb_solr_query}}}

\item{synonyms}{whether to compare with the synonym field rather than the
canonical label.}

\item{fixed}{Whether to insist on one exact match for each input identifier
in \code{ids}. When \code{fixed=FALSE} solr wildcards can be used
(returning an arbitrary number of identifiers) but query ids cannot be
matched to returned ids.}
}
\value{
A character vector of vfb ids (for \code{vfb_fromvfbids}) or external
  identifiers (for \code{vfb_tovfbids}) in the same order as the input query.
  Missing values will be denoted by \code{NA}s when \code{fixed=TRUE}. When
  \code{fixed=FALSE} and a wild-card search returns no results a character
  vector of length 0 will be returned.
}
\description{
Convert between VFB and external identifiers
}
\details{
Note that these functions are not appropriate for fetching more than
  ~ 200 identifiers individually specified identifiers. If a large number of
  identifiers are required, it is a much better idea to fetch all ids and
  then filter locally.
}
\examples{
# some flycircuit ids
fcids=c("VGlut-F-000304", "VGlut-F-200278", "fru-F-200121", "TH-F-300016")
vfbids=vfb_tovfbids(fcids)
vfb_fromvfbids(vfbids)

# make up a fake id for testing - will give a warning
\donttest{
vfb_fromvfbids(c(vfbids, "VFB_10013392"))
}
## FlyCircuit ids
# these "gene_name" ids are a secondary id on FlyCircuit
fcgns=c("DvGlutMARCM-F1106_seg1", "DvGlutMARCM-F1448_seg1",
  "FruMARCM-F001735_seg002", "THMARCM-160F_seg2")

# note the difference in order between fixed = TRUE/FALSE
vfb_tovfbids(fcgns, fixed=TRUE)
vfb_tovfbids(fcgns, fixed=FALSE)
# depends on searching synonym field
vfb_tovfbids(fcgns, fixed=TRUE, synonyms=FALSE)
vfb_tovfbids(fcids, fixed=TRUE)
# these are the canonical form, so don't depend on searching synonym field
vfb_tovfbids(fcids, synonyms=FALSE)

# Some GMR GAL4 lines
gmrs=c('93D09', '87F10')
vfb_tovfbids(sprintf("GMR_\%s*", gmrs), fixed=FALSE)
# unfortunately these don't work with fixed=TRUE to define ordering
vfb_tovfbids(sprintf("GMR_\%s*", gmrs), fixed=TRUE)
}
\seealso{
\code{\link{gmr_vfbid}}, \code{\link{vfb_solr_query}}, which powers
  the underlying queries.
}
