% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfb_query.R
\name{vfb_neo4j_query}
\alias{vfb_neo4j_query}
\title{Query VFB's Neo4J graph database}
\usage{
vfb_neo4j_query(
  x,
  ...,
  path = "db/data",
  server = getOption("vfbr.server.neo4j")
)
}
\arguments{
\item{x}{A character query in Neo4J's cypher language}

\item{...}{Additional arguments passed to \code{neo2R::cypher()}}

\item{path}{The relative path on the server for the Neo4J endpoint}

\item{server}{The server's root URL}
}
\value{
A data.frame of query results
}
\description{
Query VFB's Neo4J graph database
}
\details{
Under the hood, this uses the \code{RNeo4j::cypher} function to call
  a Neo4J service running on the specified VFB server.
}
\examples{
\donttest{
# ask for all neuronal classes
nclasses=vfb_neo4j_query("MATCH (n:Neuron:Class) RETURN n.label")
nrow(nclasses)
head(nclasses)

# Find all images with an associated neuronal class
q=paste("MATCH (p:Class:Neuron)<-[:INSTANCEOF]-(i:Individual:has_image)",
  "RETURN distinct i.label, p.label, p.symbol;")
nclasses_image=vfb_neo4j_query(q)
nrow(nclasses_image)
head(nclasses_image)

# how many neuronal classes have images?
length(unique(nclasses_image$p.label))

# look at number of images for each neuronal class
# NB some of these are rather generic classes where
# a specific cell type annotation was not available.
table_by_nclass <- table(nclasses_image$p.label)
subset(as.data.frame(table_by_nclass), Freq>200)
}
}
\references{
\url{https://neo4j.com/developer/cypher-query-language/}
}
\seealso{
\code{RNeo4j::RNeo4j}, \code{RNeo4j::cypher} and
  \url{https://neo4j.com/docs/rest-docs/3.3/}

Other query: 
\code{\link{vfb_owl_query}()},
\code{\link{vfb_solr_query}()}
}
\concept{query}
