% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfb_stack.R
\name{vfb_stack_url}
\alias{vfb_stack_url}
\title{Make a URL for a set of VFB ids to be displayed in the stack browser}
\usage{
vfb_stack_url(ids, path = "/site/stacks", server = getOption("vfbr.server"))
}
\arguments{
\item{ids}{A character vector of IDs}

\item{path}{The path on the server containing the query page}

\item{server}{The base url of the server}
}
\value{
character vector to open stack browser
}
\description{
Make a URL for a set of VFB ids to be displayed in the stack browser
}
\examples{
# some gmr lines
ids=c("VFBi_00004657","VFBi_00023207","VFBi_00023120","VFBi_00022264")
template="VFBt_00100000"
u=vfb_stack_url(c(template, ids))
\dontrun{
browseURL(u)
}
}

