% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vfbr.R
\docType{package}
\name{vfbr-package}
\alias{vfbr}
\alias{vfbr-package}
\title{Programmatic access to the VFB API}
\description{
Programmatic access to the VFB API
}
\section{package options}{


You can set a default server to connect to with the option:

\code{options(vfbr.server="http://vfbsandbox.inf.ed.ac.uk")}.
}
\examples{
\dontrun{
##
}
}
\references{
See
}
\keyword{package}

