% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfbr-package.R, R/vfbr.R
\docType{package}
\name{vfbr-package}
\alias{vfbr}
\alias{vfbr-package}
\title{Programmatic Access to the virtualflybrain.org website}
\description{
Programmatic Access to the virtualflybrain.org website

Programmatic access to the VFB API
}
\section{Package Options}{
 The following options can be set to specify default
  behaviour.

  \itemize{

  \item{\code{vfbr.server}}{ URL of main VFB server}

  \item{\code{vfbr.server.gepetto}}{ URL of VFB's gepetto server for 3D
  visualisation (see \code{\link{vfb_3dbrowser_url}}.)}

  }
}

\section{package options}{


  You can set a default server to connect to with the option
  \code{vfbr.server}. This is presently set on package load as follows:
  \code{options(vfbr.server="http://vfbsandbox.inf.ed.ac.uk")} unless you
  have previously set it (e.g. in your \code{\link[base]{Rprofile}}) to a
  different value.
}
\examples{
# Show state of vfbr package options
options()[grep('^vfbr', names(options()))]
# Show state of vfbr package options
options()[grep('^vfbr', names(options()))]
\dontrun{
example(vfb_solr_query)
}
}
\references{
See \url{http://virtualflybrain.org}
}
\seealso{
\code{\link{vfb_solr_query}}

\code{\link{vfb_solr_query}}
}
\keyword{package}

